added admin role

+++ app/Http/Controllers/ArticleController.php
@@ -15,14 +15,16 @@ use Illuminate\Support\Facades\Storage;
 
 class ArticleController extends Controller
 {
-    public function __construct(){
-        // $this->middleware('auth')->except(['search','searchTag','searchCategorie']);
-        // $this->middleware('articlemake')->only('create','store');
-        // $this->middleware('articleedit')->only('update','edit','destroy');
+    public function __construct()
+    {
+            $this->middleware('webmaster')->only('show');
+            $this->middleware('auth')->except(['search', 'searchTag', 'searchCategorie']);
+            $this->middleware('articlemake')->only('create', 'store', 'index');
+            $this->middleware('articleedit')->only('update', 'edit', 'destroy');
     }
-    
-        
-    
+
+
+
     /**
      * Display a listing of the resource.
      *
@@ -30,8 +32,8 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $articles=Article::all();
-        return view('admin.article.index',compact('articles'));
+        $articles = Article::all();
+        return view('admin.article.index', compact('articles'));
     }
 
     /**
@@ -41,9 +43,9 @@ class ArticleController extends Controller
      */
     public function create()
     {
-        $tags=Tag::all();
-        $categories=Categorie::all();
-        return view('admin.article.add',compact('tags','categories'));
+        $tags = Tag::all();
+        $categories = Categorie::all();
+        return view('admin.article.add', compact('tags', 'categories'));
     }
 
     /**
@@ -55,28 +57,31 @@ class ArticleController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'titre'=>'required|string',
-            'texte'=>'required|string',
-            'image'=>'required|image',
-            'categorie'=>'required|integer',
+            'titre' => 'required|string',
+            'texte' => 'required|string',
+            'image' => 'required|image',
+            'categorie' => 'required|integer',
+            'tag.*' => 'nullable|integer',
         ]);
-        $image=Storage::disk('public')->put('',$request->image);
-        $article=new Article();
-        $article->image=$image;
-        $article->titre=$request->titre;
-        $article->texte=$request->texte;
-        $article->user_id=Auth::id();
-        $article->categorie_id=$request->categorie;
-        $article->valide=false;
+
+        $image = Storage::disk('public')->put('', $request->image);
+        $article = new Article();
+        $article->image = $image;
+        $article->titre = $request->titre;
+        $article->texte = $request->texte;
+        $article->user_id = Auth::id();
+        $article->categorie_id = $request->categorie;
+        $article->valide = false;
         $article->save();
+        
+        if ($request->has('tag')) {
 
-        foreach($request->tag as $item){
-
-            Tag::find($item->id)->articles()->attach($article->id);
-            $article->tags()->attach($item->id);
+            foreach ($request->tag as $item) {
+                Tag::find($item)->articles()->attach($article->id);
+                $article->tags()->attach($item);
+            }
         }
         return redirect()->route('article.index');
-
     }
 
     /**
@@ -87,7 +92,7 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        $article->valide=true;
+        $article->valide = true;
         $article->save();
         return redirect()->back();
     }
@@ -100,9 +105,9 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
-        $tags=Tag::all();
-        $categories=Categorie::all();
-        return view('admin.article.edit',compact('tags','categories','article'));
+        $tags = Tag::all();
+        $categories = Categorie::all();
+        return view('admin.article.edit', compact('tags', 'categories', 'article'));
     }
 
     /**
@@ -115,29 +120,34 @@ class ArticleController extends Controller
     public function update(Request $request, Article $article)
     {
         $request->validate([
-            'titre'=>'required|string',
-            'texte'=>'required|string',
-            'image'=>'somtimes|image',
-            'categorie'=>'required|integer',
+            'titre' => 'required|string',
+            'texte' => 'required|string',
+            'image' => 'somtimes|image',
+            'categorie' => 'required|integer',
+            'tag.*' => 'nullable|integer',
 
         ]);
-        if($request->hasFile('image')){
-            if(Storage::disk('public')->exists($article->image)){
+        if ($request->hasFile('image')) {
+            if (Storage::disk('public')->exists($article->image)) {
                 Storage::disk('public')->delete($article->image);
             }
+            $image = Storage::disk('public')->put('', $request->image);
+            $article->image = $image;
         }
-        $image=Storage::disk('public')->put('',$request->image);
-        $article->image=$image;
-        $article->titre=$request->titre;
-        $article->texte=$request->texte;
-        $article->user_id=Auth::id();
-        $article->categorie_id=$request->categorie;
-        $article->valide=false;
+        $article->titre = $request->titre;
+        $article->texte = $request->texte;
+        $article->user_id = Auth::id();
+        $article->categorie_id = $request->categorie;
+        $article->valide = false;
         $article->save();
-        $article->tags()->detach();
-        foreach($request->tag as $tag){
-            Tag::where('id',$tag)->first()->articles()->attach($article->id);
-            $article->tags()->attach($tag);
+        if ($request->has('tag')) {
+
+            $article->tags()->detach();
+            
+            foreach ($request->tag as $item) {
+                Tag::find($item)->articles()->attach($article->id);
+                $article->tags()->attach($item);
+            }
         }
         return redirect()->route('article.index');
     }
@@ -150,46 +160,48 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-        if(Storage::disk('public')->exists($article->image)){
+        if (Storage::disk('public')->exists($article->image)) {
             Storage::disk('public')->delete($article->image);
         }
+        $article->tags()->detach();
         $article->delete();
         return redirect()->back();
     }
-    public function search(Request $request){
+    public function search(Request $request)
+    {
 
-        $articles=Article::where('valide',true)->where('titre','LIKE','%'.$request->titre.'%')->get();
-        $accueil=Accueil::find(1);
-        $blog=Blog::find(1);
-        $footer=Footer::find(1);
-        $contact=Contact::find(1);
-        $tags=Tag::inRandomOrder()->take(9)->get();
-        $categories=Categorie::inRandomOrder()->take(6)->get();
-        return view('blog',compact('accueil','articles','tags','categories','footer','contact','blog'));
+        $articles = Article::where('valide', true)->where('titre', 'LIKE', '%' . $request->titre . '%')->paginate(3);
+        $accueil = Accueil::find(1);
+        $blog = Blog::find(1);
+        $footer = Footer::find(1);
+        $contact = Contact::find(1);
+        $tags = Tag::inRandomOrder()->take(9)->get();
+        $categories = Categorie::inRandomOrder()->take(6)->get();
+        return view('blog', compact('accueil', 'articles', 'tags', 'categories', 'footer', 'contact', 'blog'));
     }
 
     public function searchTag(Tag $tag)
     {
-        
-        $articles=$tag->articles()->where('valide',true)->paginate(3);
-        $accueil=Accueil::find(1);
-        $blog=Blog::find(1);
-        $footer=Footer::find(1);
-        $contact=Contact::find(1);
-        $tags=Tag::inRandomOrder()->take(9)->get();
-        $categories=Categorie::inRandomOrder()->take(6)->get();
-        return view('blog',compact('accueil','articles','tags','categories','footer','contact','blog'));
+
+        $articles = $tag->articles()->where('valide', true)->paginate(3);
+        $accueil = Accueil::find(1);
+        $blog = Blog::find(1);
+        $footer = Footer::find(1);
+        $contact = Contact::find(1);
+        $tags = Tag::inRandomOrder()->take(9)->get();
+        $categories = Categorie::inRandomOrder()->take(6)->get();
+        return view('blog', compact('accueil', 'articles', 'tags', 'categories', 'footer', 'contact', 'blog'));
     }
 
     public function searchCategorie(Categorie $categorie)
     {
-        $articles=Article::where('valide',true)->where('categorie_id', $categorie->id)->paginate(3);
-        $accueil=Accueil::find(1);
-        $blog=Blog::find(1);
-        $footer=Footer::find(1);
-        $contact=Contact::find(1);
-        $tags=Tag::inRandomOrder()->take(9)->get();
-        $categories=Categorie::inRandomOrder()->take(6)->get();
-        return view('blog',compact('accueil','articles','tags','categories','footer','contact','blog'));
+        $articles = Article::where('valide', true)->where('categorie_id', $categorie->id)->paginate(3);
+        $accueil = Accueil::find(1);
+        $blog = Blog::find(1);
+        $footer = Footer::find(1);
+        $contact = Contact::find(1);
+        $tags = Tag::inRandomOrder()->take(9)->get();
+        $categories = Categorie::inRandomOrder()->take(6)->get();
+        return view('blog', compact('accueil', 'articles', 'tags', 'categories', 'footer', 'contact', 'blog'));
     }
 }
